Music Genre Classification/requirements.txt
kagglehub
pandasppip 
scikit-learn
tqdm
librosa
ffmpeg-python
matplotlib
seaborn